+++ Tentacle/Client.swift
@@ -10,7 +10,6 @@ import Argo
 import Foundation
 import ReactiveSwift
 import Result
-import Ogra
 
 extension JSONSerialization {
     internal static func deserializeJSON(_ data: Data) -> Result<Any, AnyError> {
